feat(core): fix folder structure for ci-status service
