+++ packages/ci-status/src/cache/index.ts
@@ -1,0 +1,23 @@
+export class CacheManager {
+  private cache: Map<string, any> = new Map();
+  private cacheExpiry = 2 * 60 * 1000; // 2 minutes
+
+  getFromCache(key: string): any {
+    const cached = this.cache.get(key);
+    if (cached && Date.now() - cached.timestamp < this.cacheExpiry) {
+      return cached.data;
+    }
+    return null;
+  }
+
+  setCache(key: string, data: any): void {
+    this.cache.set(key, {
+      data,
+      timestamp: Date.now(),
+    });
+  }
+
+  clearCache(): void {
+    this.cache.clear();
+  }
+}
