+++ packages/ci-status/tsconfig.json
@@ -1,7 +1,6 @@
 {
   // Extends the shared Node configuration for utility/service packages
   //   "extends": "@monorepo-dashboard/tsconfig/node.json",
-
   "compilerOptions": {
     // Output compilation results to the 'dist' directory
     "outDir": "./dist",
@@ -11,14 +10,11 @@
     "sourceMap": true,
     "esModuleInterop": true,
     "downlevelIteration": true
-
     // Since your package.json defines "main": "index.ts",
     // we set the 'root' to the package root and 'outDir' to 'dist'.
   },
-
   // Include the main entry point and any other source files
-  "include": ["index.ts", "src/**/*"],
-
+  "include": ["src/**/*"],
   // Exclude node_modules and the output directory
   "exclude": ["node_modules", "dist", "**/*.test.ts"]
 }
